% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BR_strsplit.R
\name{BR_strsplit}
\alias{BR_strsplit}
\title{Title}
\usage{
BR_strsplit(x, split)
}
\arguments{
\item{x}{一个字符串}

\item{split}{根据特定字符对字符串进行分割}
}
\value{
一个字符串向量
}
\description{
Title
}
\examples{
x <- c("alfa,bravo,charlie,delta")
BR_strsplit(x,split = ",")
}
